swap contact <-> contact_2 contents
